# apispec.yaml
openapi: '3.0.3'
info:
    title: Event Planner
    version: v1.0.0
    description: Event management system API, written in Golang.
servers:
    - url: 'http://34.238.121.143:8000'
      description: Production
...
components:
    securitySchemes:
        JWTAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Use token for authorization'
    schemas:
...
        user-response:
            type: object
            properties:
                id:
                    type: integer
                    example: 7
                name:
                    type: string
                    example: John
                email:
                    type: string
                    example: john@mail.com
                gender:
                    type: string
                    example: male
                address:
                    type: string
                    example: Groove st. Ave, Dallas
                avatar:
                    type: string
                    example: https://source.unsplash.com/1200x400/?sea
                created_at:
                    type: string
                    example: 2012/07/28T10:02:37.000+0700
                updated_at:
                    type: string
                    example: 2012/07/28T10:02:37.000+0700
...
        category-response:
            type: object
            properties:
                id:
                    type: integer
                    example: 7
                title:
                    type: string
                    example: Some random category
                created_at:
                    type: string
                    example: 2012/07/28T10:02:37.000+0700
                updated_at:
                    type: string
                    example: 2012/07/28T10:02:37.000+0700
        event-request:
            type: object
            properties:
                title:
                    type: string
                    example: 'Event Title'
                hosted_by:
                    type: string
                    example: 'Someone'
                cover:
                    type: string
                    example: 'https://source.unsplash.com/1200x400/?festival'
                datetime_event:
                    type: string
                    example: '2012-07-28T08:37:12:000+12:00'
                category_id:
                    type: integer
                    example: 3
                description:
                    type: string
                    example: Lorem ipsum dolor sit, amet consectetur adipisicing elit. Enim, inventore deserunt! Ad optio esse libero. Debitis accusantium ut labore quaerat!
        event-response:
            type: object
            properties:
                id:
                    type: integer
                    example: 7
                title:
                    type: string
                    example: 'Event title'
                hosted_by:
                    type: string
                    example: 'Someone'
                cover:
                    type: string
                    example: 'https://source.unsplash.com/1200x400/?festival'
                datetime_event:
                    type: string
                    example: '2012-07-28T08:37:12:000+12:00'
                description:
                    type: string
                    example: Lorem ipsum dolor sit, amet consectetur adipisicing elit. Enim, inventore deserunt! Ad optio esse libero. Debitis accusantium ut labore quaerat!
                user_id:
                    type: integer
                    example: 2
                user:
                    $ref: '#/components/schemas/user-response'
                category_id:
                    type: integer
                    example: 3
                category:
                    $ref: '#/components/schemas/category-response'
                created_at:
                    type: string
                    example: '2012-07-28T08:37:12:000+12:00'
                updated_at:
                    type: string
                    example: '2012-07-28T08:37:12:000+12:00'
...
        pagination:
            type: object
            properties:
                page:
                    type: integer
                    example: 1
                limit:
                    type: integer
                    example: 50
                total_pages:
                    type: integer
                    example: 11
paths:
...
    /api/events:
        get:
            tags:
                - 'Event'
            description: Getting list of events
            summary: getting list of events
            parameters:
                - in: query
                  name: q
                  allowEmptyValue: true
                  description: Search by events title and description
                  schema:
                      type: string
                      example: Some event's name to search
                - in: query
                  name: location
                  description: Search events by location
                  schema:
                      type: string
                      example: Surabaya
                - in: query
                  name: limit
                  description: Limit rows per page
                  schema:
                      type: string
                      example: 50
                - in: query
                  name: page
                  description: Page number
                  schema:
                      type: string
                      example: 1
            responses:
                '200':
                    description: Success getting list of user
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        example: OK
                                    error:
                                        type: string
                                        example: null
                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                example: 'http://example.com/api/events'
                                            next:
                                                type: string
                                                example: 'http://example.com/api/events?p=2'
                                            last:
                                                type: string
                                                example: 'http://example.com/api/events?p=10'
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/event-response'
                                    pagination:
                                        $ref: '#/components/schemas/pagination'
                '500':
                    description: server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        example: 'ERROR'
                                    code:
                                        type: string
                                        example: 500
                                    error:
                                        type: string
                                        example: 'server error, cannot connect to database'

                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                example: 'http://example.com/api/events'

        post:
            tags:
                - 'Event'
            security:
                - JWTAuth: []
            description: Create single event
            summary: Create single event
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/event-request'
            responses:
                '200':
                    description: success creating single user
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        example: OK
                                    code:
                                        type: integer
                                        example: 201
                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                example: 'http://example.com/api/users/2'
                                    data:
                                        $ref: '#/components/schemas/user-response'
                '400':
                    description: validation failed
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        example: 'ERROR'
                                    code:
                                        type: string
                                        example: 400
                                    error:
                                        type: string
                                        example: 'validation error'
                                    errors:
                                        type: array
                                        items:
                                            type: object
                                            properties:
                                                field:
                                                    type: string
                                                    example: username
                                                error:
                                                    type: string
                                                    example: required

                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                example: 'http://domain.com/api/users'
                '500':
                    description: server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        example: 'ERROR'
                                    code:
                                        type: string
                                        example: 500
                                    error:
                                        type: string
                                        example: 'server error, cannot connect to database'

                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                example: 'http://example.com/api/events'
...